Add analytics ttl settings to Config spec to allow configuration (#334)
* Added analytics ttl settings to config spec to allow for configuration

* Added and modified unit tests. Added descriptions of the new fields in
crds.

* Added explicit ttl settings to openshift manager manifest

* Lowered ttl's in openshift manifest
