+++ deploy/openshift/manifests/0000000-contrail-09-manager.yaml
@@ -41,6 +41,10 @@ spec:
             node-role.kubernetes.io/master: ""
         serviceConfiguration:
           cassandraInstance: cassandra1
+          analyticsDataTTL: 2
+          analyticsConfigAuditTTL: 2160
+          analyticsFlowTTL: 1
+          analyticsStatisticsTTL: 2
           containers:
             - name: analyticsapi
               image: <CONTRAIL_REGISTRY>/contrail-analytics-api:<CONTRAIL_VERSION>-rhel
