+++ pkg/apis/contrail/v1alpha1/tests/BUILD.bazel
@@ -3,6 +3,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_test")
 go_test(
     name = "go_default_test",
     srcs = [
+        "config_config_test.go",
+        "config_customized_test.go",
         "config_test.go",
         "config_vrouter_test.go",
     ],
